Move partial into separate file + shush clippy
